requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.0.0
jsonpatch==1.28
jsonpointer==2.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-python==3.4.10
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==20.0.0
requests==2.25.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.4.0
torchfile==0.1.0
torchvision==0.5.0
tornado==6.1
traitlets==4.3.3
urllib3==1.26.2
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==0.57.0
